refactor naming of projection types
