+++ src/Elastic.CommonSchema.Generator/Views/IndexTemplates.Generated.cshtml
@@ -1,7 +1,7 @@
 @using System
 @using System.Linq
 @using Generator
-@inherits Elastic.CommonSchema.Generator.Views.CodeTemplatePage<Elastic.CommonSchema.Generator.Domain.CsharpProjection>
+@inherits Elastic.CommonSchema.Generator.Views.CodeTemplatePage<Elastic.CommonSchema.Generator.Projection.CommonSchemaTypesProjection>
 // Licensed to Elasticsearch B.V under one or more agreements.
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
